Struct dsp::VisitOrderReverse

pub struct VisitOrderReverse {
    // some fields omitted
}

A walker type for walking over a Graph's nodes in the order in which they will visited when audio is requested from the Graph.

Methods

impl VisitOrderReverse

fn next<F, N>(&mut self, graph: &Graph<F, N>) -> Option<NodeIndex>

The index of the next node that would be visited during audio requested in our walk of the given Graph's visit order.