Trait dsp::FromSample

pub trait FromSample<S> {
    fn from_sample_(s: S) -> Self;
}

Similar to the std From trait, but specifically for converting between sample types.

We use this trait to be generic over the Sample::to_sample and Sample::from_sample methods.

Required Methods

fn from_sample_(s: S) -> Self